requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
compressai
pandas
tqdm
Pillow